% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingscoreAMLMutations.R
\docType{package}
\name{SingscoreAMLMutations-package}
\alias{SingscoreAMLMutations}
\alias{SingscoreAMLMutations-package}
\title{SingscoreAMLMutations: Using singscore to predict mutations in AML from transcriptomic signatures}
\description{
This workflow package shows how transcriptomic signatures can be used
  to infer phenotypes. The workflow begins by showing how the TCGA AML transcriptomic
  data can be downloaded and processed using the TCGAbiolinks packages. It then shows
  how samples can be scored using the singscore package and signatures from the 
  MSigDB. Finally, the predictive capacity of scores in the context of predicting
  a specific mutation in AML is shown.The workflow exhibits the interplay of
  Bioconductor packages to achieve a gene-set level analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisLaboratory/SingscoreAMLMutations}
  \item Report bugs at \url{https://github.com/DavisLaboratory/SingscoreAMLMutations/issues}
}

}
\author{
\strong{Maintainer}: Dharmesh D. Bhuva \email{bhuva.d@wehi.edu.au} (\href{https://orcid.org/0000-0002-6398-9157}{ORCID})

Authors:
\itemize{
  \item Momeneh Foroutan \email{foroutan.m@wehi.edu.au} (\href{https://orcid.org/0000-0002-1440-0457}{ORCID})
  \item Yi Xie \email{xie.y@wehi.edu.au} (\href{https://orcid.org/0000-0002-1938-4089}{ORCID})
  \item Ruqian Lyu \email{lyu.r@wehi.edu.au}
  \item Joseph Cursons \email{cursons.j@wehi.edu.au} (\href{https://orcid.org/0000-0002-5053-4540}{ORCID})
  \item Melissa J. Davis \email{davis.m@wehi.edu.au} (\href{https://orcid.org/0000-0003-4864-7033}{ORCID})
}

}
